Push the passport strategy creators out into their own modules
