+++ middleware/passport/aadStrategy.ts
@@ -1,0 +1,79 @@
+import { IProviders } from '../../transitional';
+
+import { OIDCStrategy } from 'passport-azure-ad';
+
+function activeDirectorySubset(app, iss, sub, profile, done) {
+  // CONSIDER: TODO: Hybrid tenant checks.
+  // Internal-only code:
+  // ----------------------------------------------------------------
+  // We've identified users with e-mail addresses in AAD similar to
+  // myoutlookaddress#live.com. These are where people have had work
+  // shared with them through a service like Office 365; these users
+  // are not technically employees with active credentials, and so
+  // they should *not* have access. We reject here before the
+  // session tokens can be saved.
+  // if (username && username.indexOf && username.indexOf('#') >= 0) {
+  //   return next(new Error('Your hybrid tenant account, ' + username + ', is not permitted for this resource. Were you invited as an outside collaborator by accident? Please contact us if you have any questions.'));
+  // }
+
+  const config = app.settings.runtimeConfig;
+  const providers = app.settings.providers as IProviders;
+  if (config && config.impersonation && config.impersonation.corporateId) {
+    const impersonationCorporateId = config.impersonation.corporateId;
+    return providers.graphProvider.getUserById(impersonationCorporateId, (err, impersonationResult) => {
+      if (err) {
+        return done(err);
+      }
+      console.warn(`IMPERSONATION: id=${impersonationResult.id} upn=${impersonationResult.userPrincipalName} name=${impersonationResult.displayName}`);
+      return done(null, {
+        azure: {
+          displayName: impersonationResult.displayName,
+          oid: impersonationResult.id,
+          username: impersonationResult.userPrincipalName,
+        },
+      });
+    });
+  }
+  const subset = {
+    azure: {
+      displayName: profile.displayName,
+      oid: profile.oid,
+      username: profile.upn,
+    },
+  };
+  return done(null, subset);
+}
+
+export default function createAADStrategy(app, config) {
+  let aadStrategy = new OIDCStrategy({
+    redirectUrl: config.activeDirectory.redirectUrl || `${config.webServer.baseUrl}/auth/azure/callback`,
+    allowHttpForRedirectUrl: config.containers.docker || config.webServer.allowHttp,
+    // @ts-ignore
+    realm: config.activeDirectory.tenantId,
+    clientID: config.activeDirectory.clientId,
+    clientSecret: config.activeDirectory.clientSecret,
+    identityMetadata: 'https://login.microsoftonline.com/' + config.activeDirectory.tenantId + '/.well-known/openid-configuration',
+    responseType: 'id_token code',
+    responseMode: 'form_post',
+    // oidcIssuer: config.activeDirectory.issuer,
+    // validateIssuer: true,
+  }, activeDirectorySubset.bind(null, app));
+
+  // Patching the AAD strategy to intercept a specific state failure message and instead
+  // of providing a generic failure message, redirecting (HTTP GET) to the callback page
+  // where we can offer a more useful message
+  // @ts-ignore
+  const originalFailWithLog = aadStrategy.failWithLog;
+  // @ts-ignore
+  aadStrategy.failWithLog = function () {
+    const args = Array.prototype.slice.call(arguments);
+    const messageToIntercept = 'In collectInfoFromReq: invalid state received in the request';
+    if (args.length === 1 && typeof (args[0]) === 'string' && args[0] === messageToIntercept) {
+      return this.redirect('/auth/azure/callback?failure=invalid');
+    } else if (args.length === 1 && typeof (args[0]) === 'string') {
+      console.warn(`AAD Failure: ${args[0]}`);
+    }
+    originalFailWithLog.call(this, args);
+  };
+  return { 'azure-active-directory': aadStrategy };
+}
